requirements.txt
prisma
requests
python-dotenv
